+++ lib/uuidx/gem_version.rb
@@ -2,5 +2,5 @@
 
 module Uuidx
   # The gem version.
-  VERSION = "0.9.0"
+  VERSION = "0.10.0"
 end